+++ packages/orq-rc/RELEASES.md
@@ -5209,3 +5209,13 @@ Based on:
 - [typescript v4.0.0-rc.55] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.55] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.55 - packages/orq-rc
+
+## 2025-11-20 11:17:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.658.2 (2.755.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.56] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.56] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.56 - packages/orq-rc